Split Command and Hypervisor classes from golem.docker.manager
